server/requirements.txt
Flask==3.1.2
flask_cors==6.0.1
flask_limiter==3.12
gunicorn==23.0.0
openai==1.101.0
pydantic==2.11.7
pytest==8.4.1
python-dotenv==1.1.1
